Fix production bugs surfaced by first real integration run; run integration tests in CI
Bugs found by standing up Postgres locally (these tests had never executed):
- pg returned DATE columns as JS Date objects while all code treats dates as
  YYYY-MM-DD strings: prediction submission window never opened (always 409
  PREDICTION_CLOSED) and GET /api/predictions/current 500'd on .split of a
  Date. Fixed with a global DATE type parser in db/pool.ts.
- 002 seeds only 18 bingo challenges but generateCard requires 24 distinct →
  card creation always threw. New migration 003 adds 8 more (pool now 26).
- closeWeekPredictions reused $2 in two type contexts → 'inconsistent types
  deduced for parameter $2'. Explicit ::int casts.
- Prediction integration tests were Monday-dependent (window = Monday only);
  tests now move starts_on to today to open the window deterministically.

Test infra: shared resetDatabase helper applies every migration in order;
vitest fileParallelism off (suites share one DB); CI gets a postgres:16
service so all 43 tests run on every push.

+++ apps/api/src/services/weekClose.ts
@@ -51,11 +51,13 @@ export async function closeWeekPredictions(
     );
   }
 
+  // $2 appears in two type contexts (assignment + comparison) — the explicit
+  // casts keep Postgres from deducing inconsistent parameter types
   await db.query(
     `UPDATE weeks
      SET status = 'closed',
-         group_total_steps = $2,
-         target_hit = ($2 >= $3)
+         group_total_steps = $2::int,
+         target_hit = ($2::int >= $3::int)
      WHERE id = $1`,
     [weekId, actualTotal, group_target_steps],
   );
